+++ src/Auth/CanDirective.php
@@ -157,7 +157,6 @@ GRAPHQL;
                     $root,
                     $args,
                     $context,
-                    $resolveInfo,
                 )
                 ->get();
         }
@@ -194,7 +193,6 @@ GRAPHQL;
                     $root,
                     $args,
                     $context,
-                    $resolveInfo,
                     Utils::instanceofMatcher(TrashedDirective::class),
                 );
                 assert($enhancedBuilder instanceof EloquentBuilder);