+++ src/Schema/Directives/AllDirective.php
@@ -5,13 +5,10 @@ namespace Nuwave\Lighthouse\Schema\Directives;
 use GraphQL\Language\AST\FieldDefinitionNode;
 use GraphQL\Language\AST\InterfaceTypeDefinitionNode;
 use GraphQL\Language\AST\ObjectTypeDefinitionNode;
-use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
-use Illuminate\Database\Eloquent\Relations\Relation;
-use Illuminate\Database\Query\Builder as QueryBuilder;
 use Illuminate\Support\Collection;
-use Laravel\Scout\Builder as ScoutBuilder;
 use Nuwave\Lighthouse\Execution\ResolveInfo;
 use Nuwave\Lighthouse\Schema\AST\DocumentAST;
+use Nuwave\Lighthouse\Schema\Directives\Traits\HasBuilderArgument;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Support\Contracts\FieldManipulator;
 use Nuwave\Lighthouse\Support\Contracts\FieldResolver;
@@ -19,6 +16,8 @@ use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 
 class AllDirective extends BaseDirective implements FieldResolver, FieldManipulator
 {
+    use HasBuilderArgument;
+
     public static function definition(): string
     {
         return /** @lang GraphQL */ <<<'GRAPHQL'
@@ -51,30 +50,15 @@ GRAPHQL;
 
     public function resolveField(FieldValue $fieldValue): callable
     {
-        return function (mixed $root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo): Collection {
-            if ($this->directiveHasArgument('builder')) {
-                $builderResolver = $this->getResolverFromArgument('builder');
-
-                $query = $builderResolver($root, $args, $context, $resolveInfo);
-                assert(
-                    $query instanceof QueryBuilder || $query instanceof EloquentBuilder || $query instanceof ScoutBuilder || $query instanceof Relation,
-                    "The method referenced by the builder argument of the @{$this->name()} directive on {$this->nodeName()} must return a Builder or Relation.",
-                );
-            } else {
-                $query = $this->getModelClass()::query();
-            }
-
-            return $resolveInfo
-                ->enhanceBuilder(
-                    $query,
-                    $this->directiveArgValue('scopes', []),
-                    $root,
-                    $args,
-                    $context,
-                    $resolveInfo,
-                )
-                ->get();
-        };
+        return fn (mixed $root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo): Collection => $resolveInfo
+            ->enhanceBuilder(
+                $this->makeBuilder($root, $args, $context, $resolveInfo),
+                $this->directiveArgValue('scopes', []),
+                $root,
+                $args,
+                $context,
+            )
+            ->get();
     }
 
     public function manipulateFieldDefinition(DocumentAST &$documentAST, FieldDefinitionNode &$fieldDefinition, ObjectTypeDefinitionNode|InterfaceTypeDefinitionNode &$parentType): void