+++ CHANGELOG.md
@@ -9,6 +9,12 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+## v6.9.2
+
+### Fixed
+
+- Fix `DirectiveLocator` assertion
+
 ## v6.9.1
 
 ### Fixed